Updated description of git push command
Revised the content related to the git push command, applying company style.

Rephrased so we're stating that the git push command is what's performing the action.

Rewrote the content to be present tense and contain directive language (telling the user what to do based on a scenario).

I removed some content that seemed self-evident.  Specifically, I pulled out the statement about how this is used to share code with a remote branch.  The basic description seems to cover that already.