Revised git fetch description
Updated content for the git fetch command.

Spelled out "repo" as "repository".  Revised language to be all in present tense and directive, per the style guide.

I removed most of the references to 'master", except for the 'git merge origin/master' command.  There are audiences who might react to "master" as non-inclusive language.