Updated description of the git pull command
Revised the description of the git pull command to make the content display in present tense and active voice. Also rephrased to be directive, where appropriate.

Revised to remove 'under the hood', which is a figure of speech and might confuse some readers.

Are 'git push' and 'git pull' described as equivalent or do we mean that 'git fetch' and 'git pull' are equivalent?   I thought we might mean 'fetch' and 'pull' are equivalent and I revised the text based on that.  Please let me know if I misunderstood.